camera: Add initial FPS camera implementation, integrate with engine
The camera struct keeps track of its' position and orientation, which
are updated in response to user events through mutator functions.
The camera keeps the timestamp of the last frame, in order to
provide framerate-independent motion. So far, only motion with the
WASD keys is supported, not changing orientation with the mouse.