Blender export script: Various fixes
The first thing fixed is the coordinate systems of both the vertices
and the bones. There were inconsitencies between them, and the
script did not take into account parenting relationships.

As well, the script has been amended to only export selected objects,
meaning it is necessary to select the armatures and meshes before
exporting. This is much more reliable and the script doesn't get
broken by a random stray object in the scene.