render, anim: Rendering of skeleton in different frames
So, this is a pretty gigantic commit. It took some time to get this
to work.

The PFOBJ file format was updated, the engine-side parser
had to be updated as well. The new format includes a
transform to the parent bone's space for each joint in the bind pose.
This means that we need to walk up the joint heirarchy when building
our final bind pose matrix. The animation clip data now includes a
local transformation for each joint in the current frame. To build a
skeleton for the current frame, we also traverse up the bone heirarchy,
combining the local transformation with the transform to the parent
bone's space at each step.

New APIs were added to the 'anim' subsystem for getting the skeleton,
posed for the current frame. Utility functions needed for the feature,
as well as some necessary refactoring also made it in as part of this
commit.