Blender export script: Update format of bind pose and frame data
Specifically, the bind pose will not be exported as a transformation
to its' parent joint's coordinate space. In the case of a root bone,
the transformation will take the bone to world space. Each keyframe
will also hold a local transformation (in the current joint's coordinate
space) for each joint.

The transformations are stored as a 3-tuple of scale, rotation, and
translation. Scale and translation are 3-vectors. Rotation is stored
in Euler angles in XYZ order. Within the engine, the rotation will
actually be stored as a quaternion, but there is some trickiness with
exporting quaternions from Blender to the engine that still need to
be figured out.