Implement GPU vertex skinning
An array of the inverse bind matrices and another array of the
current pose matrices will be sent to the vertex shader in the
form of a uniform. This will be done for each entity/mesh. In
the shader, the matrices will be used to compute the final
position of the vertex, using the veretx's 'joint_indices' and
'joint_weights' attributes.

Some refactoring and cleanup weas done to make this possible.