render, anim: Slight refactor of where relevant shaders for a uniform are
Before, we had passed the shader names right to the render subsystem.
If we set the same uniform for a few shaders programs (ex. view matrix),
we would need to make 3 separate calls. This became a hassle to
maintain. So all relevant shaders for a uniform are now localized in
tables within render subsystem.