Implement CHUNK_RENDER_MODE_PREBAKED
This will render the entire top surface of the map to a texture using
the terrain shader and that texture will be used for coloring the
top surface using a basic shader. This is much faster since we are
doing basic texturing and do not need to do real-time blending. As
well, all non-visible vertices are stripped from the chunk mesh in
this mode.