Implement shadow mapping.
Rendering will now happing in 2 passes. The first one will render to the
shadow map from the light's point of view and the second one will sample
the shadow map to determine if fragments are in shadow or not.

Some slight cleanup, such as adding 'GL_ASSERT_OK', was done during the
implementation of this feature.