nuklear: add support for virtual window resolutions
Every window gets its' own 'virtual resolution', which is what
it believes the physical window resolution is. The window's
coordinates are then mapped to the physical resolution. This allows
coordinates within a window to be specified in terms of its' own
virtual resolution, and the coordinates will be scaled appropriately
for the physical resolution. This enables scaling of UI for different
resolutions, which nuklear could not handle before as it operated
on just pixels.