Movement: use weighted moving average to smooth entity's turning
A weighted average of the entity's past N velocities is used for
setting the orientation of moving entites. This smooths the
turning of the entities, curbing the toggling or oscillatory
behaviour that could arise before. This also makes the entity
a bit slower and more sluggish in turning but, overall, this
makes the movement behaviour smoother and more natural-looking
than before.