Implement 'ememy seek fields' for efficient target finding during combat:
This allows much better scaling for large groups and thre is
no longer a massive frame lag when 100+ units enter combat.

With the new system, entities will share a field that will
guide them towards enemies of the opposite faction. This way,
the will not keep on requesting A* grid paths, which are very
expensive to compute. As well, all entities on a chunk of a
single faction re-use the same enemy seek field.

Lots of other drive-by cleanup got into this patch that was
added in the process of supporting this feature.