nuklear: Introduce means to identify a texture resource by a string and defer lookup/loading until command conversion time
This change is in preparation for moving all the rendering code to a separate
thread. In this case, we wanted to hide the details of the OpenGL texture ID
from the UI scripting code, as well as to move all the texture-related
OpenGL calls (such as allocating an ID) to rendering-time.