Multithread Permafrost Engine
Split up the code into a rendering and a simulation thread. The
simulation thread pushes commands to a queue and the render
thread executes the draw calls for the previous frame's commands.

Any state necessary for rendering is copied to a stack. So the
render thread has a copy of any previous frame state that it needs.
As such, no synchronization is required between the threads,
except a swap of the command queue pointers.

Existing R_GL calls that would call the OpenGL APIs directly
are now converted to deferred calls which just push a command and
arguments to the render thread, to be invoked from there during
the next tick.