pickling: properly pickle/unpickle subclasses of builtins
In this new scheme, types 'tp_new' and 'tp_init' is invoked
not via some C API like 'PyList_New' but via calling a type
such as PyList_Type. When needing to create instances of builtin
subclasses, we use a dummy subclass that is a direct descendant
of the builtin for construction.