stalloc.c: Make sure returned pointer is aligned to sizeof(intmax_t), and zero out pad bytes
The alignment is expected by some routines (ex. nuklear vertex buffer allocation).
The zero-padding is for when we copy an unaligned size to the returned buffer and
then dereference it. Not sure how well-defined/portable this is, but it works.

This commit fixes the UI flickering bug that happened every time a model was loaded
(it was due to nuklear vertex buffer returned from stalloc ending up unaligned).