Introduce subsessions and the subsession stack:
A subsession is the current engine state (what was formerly a session).
A session is a stack of such subsessions. Using an API, it is possible
to push a new subsession, effectively using it to replace the old
subsession, while also pushing the old subsession onto an internal
stack. Later, it is possible to pop a subsession.

This allow more easily organizing the engine code as a set of independent
scripts between which control can be passed. (ex. the main menu script
can be separate from a mission script)