Make the OpenGL context in the main thread, then hand it off to the render thread
Do this instead of the old approach of making the OpenGL context in the render
thread directly. The reason is was observed that with the old approach
glewInit would fail on some systems (Nvidia card on Linux for example).

Thanks to @draczer01 (https://github.com/draczer01) for helping debug the issue
and test out a fix.