CPython: Add per-opcode tracing hook
During frame evaluation, CPython NULLs out the current frame's
f_stacktop member. As such, there is no reliable way to get
the size of the current evaluation stack of a frame. One exception
is that CPython sets f_stacktop to the correct value for trace
calls. In the current version of CPython, the tracing is available
only on a per-line basis. So we add this hook to be able to grab
the value of f_stacktop before every single opcode execution. This
allows us to reliably know the size of the evaluation stack before
every opcode.

Knowing the size of the evaluation stack is necessary to reliably
save and restore frame objects that are partially through their
execution.