Render translucent objects last
When they are rendered before non-translucent objects, they fill the
depth buffer, meaning that later rendered objects will not be visible
when they should be.

In the current version of the code, there can still be incorrect cases
when animated objects are behind non-animated objects. (due to the
fact that all animated objects get rendered before all static objects)
but this is sufficient for now.