Improve behaviour of 'enemy seek fields':
A region larger than the size of the chunk is used for computing
the integration field for enemy seek fields. Then the center of this
region is used as the enemy seek field.

This allows the enemy seek fields to also include enemies outside
the boundaries of the chunk as valid targets and have the enemies
attack more 'seamlessly'.