Fix wild pointer dereference in entity.c:
We were giving a UID (key) instead of an iterator (bin index)
to kh_del, resulting in overwriting a word of memory not owned
by the hash table, but likely still writable in the process address
space. The result was some fun, fun times where a random word
of memory (in CPython for example) would get 0x5s or 0x0s written
to it, causing (probably) a crash at whatever time it finally got
discovered. Shoutout to ASan for the debug help.