navigation: Improve portal graph pathfinding to address some problematic cases
Instead of just finding a series of portals to go through, we now find a series
of portals along with the local island ID on which we enter the chunk through
the portal. Some comments with ASCII diagrams are added to the code to show why
this is necessary. Multiple cases (specifically where a unit needs to travel
through the same chunk, or even the same portal, more than once to reach his
optimal target) are now working better as a result of this patch.